New language: agreement!
The students were confused a bit about the distinction between compiling and
interpreting. There is a lot of ways to address this confusion, making it about
types ("Prog -> Val" vs "Prog -> Prog"), for example.

Here we are introducing a new language that builds on abscond by introducing a
primitive operation that gets an int from the user.