Getting asm-interp and Extort to play nice.
There's a challenge in getting Extort and asm-interp to play nice
together.  The problem is that Extort, which adds handled errors,
calls a C function named `error` which prints a symbol and exits.  If
you do this from `asm-interp` this will exit whatever is running
`asm-interp`, prematurely exiting out of whatever you're trying to do
(build the web page, test the code, etc.).  Not what you want.

The solution, as always, is a level of indirection.  The run-time now
has a function pointer called `error`.  The `main` procedure sets to
something that prints and exits.  On the other hand `asm-interp` sets
it to something that raises an exception and then handles it.

This commit isn't the end-all for these kinds of issues, but it's a
first cut and the API for hooking in to the run-time can be refined
based on this basic idea.