Avoid boolean blindness in compile-time stack
In carrying around a CEnv, we were using `#f` to
represent non-variable values being put on the stack.

So when a multi-argument primitive places one of its subexpressions on the
stack we would do `(cons #f c)` in order to illustrate that there was space
taken up on the stack, but it should never be the result of a `lookup`.

This works great for Fraud and Hustle, but when we get to Iniquity, Jig, Loot,
Shakedown (anything with function calls that needs to adhere to calling
conventions) we also use `#f` to represent padding for alignment _and_ for the
PC that's placed on the stack for a call. This means that we can not statically
determine the stack layout. Making non-trivial stack re-arrangement difficult.

This commit ads a notion of 'Slots' that we can use in order to be more precise
about _what_ is taking up a stack slot. For Fraud and Hustle (this commit) the
change is trivial to the point of unnecessary, but it is my belief that
starting to use this mechanism here (even though it's equivalent to just using
`#f`) is the correct pedagogical approach.